feat: 분산 투자 비율 통계 반영, 테스트 코드 작성

+++ BE/mmmmBatch/src/main/java/com/beautifulyomin/mmmmbatch/batch/analysis/entity/InvestmentReport.java
@@ -46,7 +46,7 @@ public class InvestmentReport {
 
     @Column(precision = 5, scale = 2)
     @Range(min = 0, max = 100)
-    private BigDecimal diversification;
+    private Integer diversification;
 
     @Column(nullable = false)
     @Range(min = 0, max = 100)
